requirements.txt
python-telegram-bot
cloudscraper
beautifulsoup4
requests